requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
mysql-connector-python==8.4.0
requests==2.32.3
urllib3==2.2.1
